login for new profile experience
- use a singleton for the mysql wrapper to simplify depending on it
- collect all user_clients for a given client_id and email
- remove old admin login
- move demographic data options into separate file
