+++ web/src/components/pages/profile/demographics/info.tsx
@@ -1,0 +1,8 @@
+import * as React from 'react';
+import Layout from '../layout';
+
+export default class ProfilePage extends React.Component {
+  render() {
+    return <Layout>wat</Layout>;
+  }
+}
