+++ web/src/components/pages/profile/info/info.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+
+export default class ProfilePage extends React.Component {
+  render() {
+    return <div>wat</div>;
+  }
+}
