+++ web/src/components/profile-form/profile-form.tsx
@@ -7,9 +7,10 @@ import pick = require('lodash.pick');
 import * as React from 'react';
 import { connect } from 'react-redux';
 import { DEFAULT_LOCALE } from '../../services/localization';
+import { ACCENTS, AGES, GENDERS } from '../../stores/demographics';
 import { Locale } from '../../stores/locale';
 import StateTree from '../../stores/tree';
-import { ACCENTS, AGES, GENDERS, User } from '../../stores/user';
+import { User } from '../../stores/user';
 import Modal from '../modal/modal';
 import { Button, Hr, LabeledInput, LabeledSelect, TextButton } from '../ui/ui';
 import { isContributable } from '../locale-helpers';
